LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/implicit-submission.optional-expected.txt





FAIL Submit event with a submit button assert_equals: expected (object) Element node <input name="submitButton" type="submit"></input> but got (undefined) undefined
FAIL Submit event with no submit button assert_equals: expected (object) null but got (undefined) undefined
